Updated build system

+++ Rakefile
@@ -19,7 +19,9 @@ task :setup do
     if File.exist? "/usr/local/bin/wget"
       puts "wget is already installed"
     else
-      system("curl -a -O http://ftp.gnu.org/gnu/wget/wget-1.11.4.tar.bz2")
+      if !File.exist? "wget-1.11.4.tar.bz2"
+        system("curl -a -O http://ftp.gnu.org/gnu/wget/wget-1.11.4.tar.bz2")
+      end
       system("bunzip2 wget-1.11.4.tar.bz2")
       system("tar xf wget-1.11.4.tar")
       Dir.chdir("wget-1.11.4") do
@@ -32,10 +34,12 @@ task :setup do
     if File.exist? "/usr/local/bin/pcregrep"
       puts "pcre is already installed"
     else
-      system("curl -a -O ftp://ftp.csx.cam.ac.uk/pub/software/programming/pcre/pcre-7.9.tar.bz2")
-      system("bunzip2 pcre-7.9.tar.bz2")
-      system("tar xf pcre-7.9.tar")
-      Dir.chdir("pcre-7.9") do
+      if !File.exist? "pcre-8.10.tar.bz2"
+        system("curl -a -O ftp://ftp.csx.cam.ac.uk/pub/software/programming/pcre/pcre-8.10.tar.bz2")
+      end
+      system("bunzip2 pcre-8.10.tar.bz2")
+      system("tar xf pcre-8.10.tar")
+      Dir.chdir("pcre-8.10") do
         system("./configure")
         system("make install")
       end
@@ -45,10 +49,12 @@ task :setup do
     if File.exist? "/opt/godi/bin/ocaml"
       puts "GODI is already installed"
     else
-      system("curl -a -O http://download.camlcity.org/download/godi-rocketboost-20090916.tar.gz")
-      system("tar xvzf godi-rocketboost-20090916.tar.gz")
+      if !File.exist? "godi-rocketboost-20091222.tar.gz"
+        system("curl -a -O http://download.camlcity.org/download/godi-rocketboost-20091222.tar.gz")
+      end
+      system("tar xvzf godi-rocketboost-20091222.tar.gz")
       mkdir_p "godi"
-      Dir.chdir("godi-rocketboost-20090916") do
+      Dir.chdir("godi-rocketboost-20091222") do
         system("./bootstrap")
         system("PATH=/opt/godi/bin:/opt/godi/sbin:$PATH && ./bootstrap_stage2")
       end
